misc/HtmlTestSuite/results/test-0012.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>SVGTextContentElement.getExtentOfChar</title>
<link href="https://svgwg.org/svg2-draft/text.html#__svg__SVGTextContentElement__getExtentOfChar" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

</head><body><svg xmlns="http://www.w3.org/2000/svg" height="600" width="800">
<text><tspan id="tspan1" x="50 150 100" y="100">abc</tspan></text>
</svg>

<script>
test(() =&gt; {
  const element = document.querySelector('#tspan1');
  assert_equals(element.getNumberOfChars(), 3);
  assert_equals(element.getExtentOfChar(0).x, 50);
  assert_equals(element.getExtentOfChar(1).x, 150);
  assert_equals(element.getExtentOfChar(2).x, 100);
}, 'Multiple chunks in a tspan');
</script>
</body></html>
